requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pyodbc==4.0.35
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
email-validator==2.0.0
alembic
python-multipart
bcrypt==4.0.1